Add a first test using GTest
Check that the rectangle position is updated after calling moveRectangle().
